fix: resolve incorrect integer type conversions flagged by code scanning
Use strconv.ParseInt with explicit bit size instead of strconv.Atoi to
avoid unsafe int-to-int32 narrowing. Return uint32 directly from
shouldSwitchUser to match syscall.Credential field types and eliminate
redundant conversions.

+++ backend/plugin/parser/elasticsearch/parser.go
@@ -899,7 +899,7 @@ func (p *parser) string() (string, error) {
 						if err != nil {
 							break
 						}
-						uffff = (uffff << 4) + int(hex)
+						uffff = (uffff << 4) + int(uint32(hex))
 					}
 					// Treat uffff as UTF-16 encoded rune.
 					s += string(rune(uffff))
